+++ lib/clixon/clixon_xml_map.h
@@ -73,5 +73,6 @@ int xml_spec_populate(cxobj *x, void *arg);
 int api_path2xpath_cvv(yang_spec *yspec, cvec *cvv, int offset, cbuf *xpath);
 int api_path2xpath(yang_spec *yspec, char *api_path, cbuf *xpath);
 int api_path2xml(char *api_path, yang_spec *yspec, cxobj *xtop, cxobj **xpathp, yang_node **ypathp);
+int xml_merge(cxobj *x0, cxobj *x1, yang_spec *yspec);
 
 #endif  /* _CLIXON_XML_MAP_H_ */