+++ lib/src/clixon_proto_client.c
@@ -476,6 +476,70 @@ clicon_rpc_unlock(clicon_handle h,
     return retval;
 }
 
+/*! Get database configuration and state data
+ * Same as clicon_proto_change just with a cvec instead of lvec
+ * @param[in]  h        CLICON handle
+ * @param[in]  xpath    XPath (or "")
+ * @param[out] xt       XML tree. Free with xml_free. 
+ *                      Either <config> or <rpc-error>. 
+ * @retval    0         OK
+ * @retval   -1         Error, fatal or xml
+ * @code
+ *    cxobj *xt = NULL;
+ *    if (clicon_rpc_get(h, "/", &xt) < 0)
+ *       err;
+ *   if ((xerr = xpath_first(xt, "/rpc-error")) != NULL){
+ *	clicon_rpc_generate_error(xerr);
+ *      err;
+ *  }
+ *    if (xt)
+ *       xml_free(xt);
+ * @endcode
+ * @see clicon_rpc_generate_error
+ */
+int
+clicon_rpc_get(clicon_handle       h, 
+	       char               *xpath,
+	       cxobj             **xt)
+{
+    int                retval = -1;
+    struct clicon_msg *msg = NULL;
+    cbuf              *cb = NULL;
+    cxobj             *xret = NULL;
+    cxobj             *xd;
+
+    if ((cb = cbuf_new()) == NULL)
+	goto done;
+    cprintf(cb, "<rpc><get>");
+    if (xpath && strlen(xpath))
+	cprintf(cb, "<filter type=\"xpath\" select=\"%s\"/>", xpath);
+    cprintf(cb, "</get></rpc>");
+    if ((msg = clicon_msg_encode("%s", cbuf_get(cb))) == NULL)
+	goto done;
+    if (clicon_rpc_msg(h, msg, &xret, NULL) < 0)
+	goto done;
+    /* Send xml error back: first check error, then ok */
+    if ((xd = xpath_first(xret, "/rpc-reply/rpc-error")) != NULL)
+	xd = xml_parent(xd); /* point to rpc-reply */
+    else if ((xd = xpath_first(xret, "/rpc-reply/data/config")) == NULL)
+	if ((xd = xml_new("config", NULL)) == NULL)
+	    goto done;
+    if (xt){
+	if (xml_rm(xd) < 0)
+	    goto done;
+	*xt = xd;
+    }
+    retval = 0;
+  done:
+    if (cb)
+	cbuf_free(cb);
+    if (xret)
+	xml_free(xret);
+    if (msg)
+	free(msg);
+    return retval;
+}
+
 /*! Close a (user) session
  * @param[in] h        CLICON handle
  */